Finalize nabaztag api and transform serverfactory into a real plugin.
First of all, the plugins can now parse the argv as well (i.e. providing additional parameters).

This removes the "server-lib" w/ ServerFactory in favor of a real http plugin using event dispatching. Doing this, the Bot can expose the emitter if required. In addition to this, the http plugin will now provide a configuration

Additionally, this finalizes the nabaztag api connector. The plugin http is required for the public path accessibility.