Several changes due internal changes.
The internal plugin "jenkins" won't create an emitter object itself any more. This must be done on behalf of main.
The internal plugin "jenkins" won't expose an object of server anymore.

The internal logging was refactored to use the emitter. Currently, the log messages have to be emit itself. The target is logger.* (i.e. logger.message).

Because of "jenkins" can provide the cause of a build now, the main will generate a message with the culprits users of a build. This affects of additional text labels.

Due the introduction of a central logger, the base plugin class provides now a method `@log(event, messages…)`.